ui(refactor): split 'superdesk-monitoring' module.
`superdesk-monitoring` becomes:
```
├── aggregate-widget
│   ├── aggregate-widget.html
│   ├── aggregate.js
│   ├── configuration.html
│   ├── thumbnail.png
│   └── thumbnail.svg
├── controllers
│   ├── AggregateCtrl.js
│   ├── MonitoringController.js
│   └── index.js
├── directives
│   ├── AggregateSettings.js
│   ├── DeskNotifications.js
│   ├── ItemActionsMenu.js
│   ├── MonitoringGroup.js
│   ├── MonitoringGroupHeader.js
│   ├── MonitoringView.js
│   ├── SortGroups.js
│   └── index.js
├── filters
│   └── index.js
├── services
│   ├── CardsService.js
│   └── index.js
├── styles
│   ├── aggregate.less
│   └── monitoring.less
├── tests
│   ├── aggregate.spec.js
│   └── monitoring.spec.js
├── views
│   ├── aggregate-settings-configuration.html
│   ├── aggregate-settings.html
│   ├── desk-notifications.html
│   ├── highlights-view.html
│   ├── item-actions-menu.html
│   ├── monitoring-group-header.html
│   ├── monitoring-group.html
│   ├── monitoring-view.html
│   ├── monitoring.html
│   ├── personal.html
│   └── spike-monitoring.html
├── config.js
└── index.js

8 directories, 36 files
```
