+++ scripts/superdesk-monitoring/controllers/index.js
@@ -1,0 +1,2 @@
+export { MonitoringController } from './MonitoringController';
+export { AggregateCtrl } from './AggregateCtrl';
