+++ scripts/superdesk-monitoring/directives/AggregateSettings.js
@@ -1,0 +1,285 @@
+AggregateSettings.$inject = ['desks', 'workspaces', 'session', 'preferencesService', 'WizardHandler', '$filter'];
+export function AggregateSettings(desks, workspaces, session, preferencesService, WizardHandler, $filter) {
+    return {
+        templateUrl: 'scripts/superdesk-monitoring/views/aggregate-settings-configuration.html',
+        scope: {
+            modalActive: '=',
+            desks: '=',
+            deskStages: '=',
+            searches: '=',
+            deskLookup: '=',
+            stageLookup: '=',
+            searchLookup: '=',
+            groups: '=',
+            editGroups: '=',
+            onclose: '&',
+            widget: '=',
+            settings: '='
+        },
+        link: function(scope, elem) {
+
+            var PREFERENCES_KEY = 'agg:view';
+            var defaultMaxItems = 10;
+
+            scope.showGlobalSavedSearches = false;
+            scope.showPrivateSavedSearches = true;
+            scope.privateSavedSearches = [];
+            scope.globalSavedSearches = [];
+
+            scope.step = {
+                current: 'desks'
+            };
+
+            desks.initialize()
+            .then(function() {
+                scope.userLookup = desks.userLookup;
+            });
+
+            scope.$watch('step.current', function(step) {
+                if (step === 'searches') {
+                    workspaces.getActiveId().then(function(activeWorkspace) {
+                        if (activeWorkspace.type === 'workspace') {
+                            scope.showPrivateSavedSearches = true;
+                        } else {
+                            scope.showGlobalSavedSearches = true;
+                            scope.showPrivateSavedSearches = false;
+                        }
+                    });
+
+                    scope.initGlobalSavedSearches();
+                    scope.initPrivateSavedSearches();
+                }
+            });
+
+            scope.closeModal = function() {
+                scope.step.current = 'desks';
+                scope.modalActive = false;
+                scope.showGlobalSavedSearches = false;
+                scope.onclose();
+            };
+
+            scope.previous = function() {
+                WizardHandler.wizard('aggregatesettings').previous();
+            };
+
+            scope.next = function() {
+                WizardHandler.wizard('aggregatesettings').next();
+            };
+
+            scope.cancel = function() {
+                scope.closeModal();
+            };
+
+            scope.setDeskInfo = function(_id) {
+                var item = scope.editGroups[_id];
+                item._id = _id;
+                item.type = 'desk';
+                item.order = 0;
+
+                var deskOutput = scope.editGroups[_id + ':output'];
+                if (deskOutput) {
+                    deskOutput.selected = item.selected;
+                }
+
+                var scheduledDeskOutput = scope.editGroups[_id + ':scheduled'];
+                if (scheduledDeskOutput) {
+                    scheduledDeskOutput.selected = item.selected;
+                }
+            };
+
+            scope.setStageInfo = function(_id) {
+                var item = scope.editGroups[_id];
+                if (!item.type) {
+                    item._id = _id;
+                    item.type = 'stage';
+                    item.max_items = defaultMaxItems;
+                    item.order = _.size(scope.editGroups);
+                }
+            };
+
+            scope.setDeskOutputInfo = function(_id) {
+                var item = scope.editGroups[_id];
+                item._id = _id;
+                item.type = 'deskOutput';
+                item.max_items = defaultMaxItems;
+                item.order = _.size(scope.editGroups);
+
+                scope.editGroups[_id] = item;
+            };
+
+            scope.setScheduledDeskOutputInfo = function(_id) {
+                var item = scope.editGroups[_id];
+                item._id = _id;
+                item.type = 'scheduledDeskOutput';
+                item.max_items = defaultMaxItems;
+                item.order = _.size(scope.editGroups);
+
+                scope.editGroups[_id] = item;
+            };
+
+            scope.setSearchInfo = function(_id) {
+                var item = scope.editGroups[_id];
+                if (!item.type) {
+                    item._id = _id;
+                    item.type = 'search';
+                    item.max_items = defaultMaxItems;
+                    item.order = _.size(scope.editGroups);
+                }
+            };
+
+            scope.setPersonalInfo = function() {
+                var item = scope.editGroups.personal;
+                if (!item.type) {
+                    item._id = 'personal';
+                    item.type = 'personal';
+                    item.max_items = defaultMaxItems;
+                    item.order = _.size(scope.editGroups);
+                }
+            };
+
+            /**
+             * Init private saved searches with all saved searches for this user
+             */
+            scope.initPrivateSavedSearches = function() {
+                var user = session.identity._id;
+                if (scope.privateSavedSearches.length > 0) {
+                    scope.privateSavedSearches.length = 0;
+                }
+                _.each($filter('sortByName')(scope.searches), function(item) {
+                    if (item.user === user && !item.is_global) {
+                        scope.privateSavedSearches.push(item);
+                    }
+                });
+            };
+
+            /**
+             * Init global saved searches with all saved searches for all users
+             * where is_global flag is true
+             */
+            scope.initGlobalSavedSearches = function() {
+                if (scope.globalSavedSearches.length > 0) {
+                    scope.globalSavedSearches.length = 0;
+                }
+                _.each($filter('sortByName')(scope.searches), function(item) {
+                    if (item.is_global) {
+                        scope.globalSavedSearches.push(item);
+                        var group = scope.editGroups[item._id];
+                        if (group && group.selected) {
+                            scope.showGlobalSavedSearches = true;
+                        }
+                    }
+                });
+            };
+
+            /**
+             * Return the list of selected groups (stages, personal or saved searches)
+             * @return {Array} list of groups
+             */
+            scope.getValues = function() {
+                var values = Object.keys(scope.editGroups).map(function(key) {
+                    return scope.editGroups[key];
+                });
+                values = _.filter(values, function(item) {
+                    if (item.type === 'desk' || !item.selected) {
+                        return false;
+                    }
+                    if (item.type === 'stage') {
+                        var stage = scope.stageLookup[item._id];
+                        return scope.editGroups[stage.desk].selected;
+                    }
+
+                    if (item.type === 'personal') {
+                        return scope.editGroups.personal.selected;
+                    }
+                    return true;
+                });
+                values = _.sortBy(values, function(item) {
+                    return item.order;
+                });
+
+                _.each(values, function(item) {
+                    if (desks.isOutputType(item.type)) {
+                        var desk_id = item._id.substring(0, item._id.indexOf(':'));
+                        item.name = desks.deskLookup[desk_id].name;
+                    }
+                });
+
+                return values;
+            };
+
+            scope.reorder = function(start, end) {
+                var values = scope.getValues();
+                if (end.index !== start.index) {
+                    values.splice(end.index, 0, values.splice(start.index, 1)[0]);
+                    _.each(values, function(item, index) {
+                        item.order = index;
+                    });
+                }
+            };
+
+            scope.save = function() {
+                var groups = [];
+                _.each(scope.getValues(), function(item, index) {
+                    if (item.selected && item.type !== 'desk') {
+                        groups.push({
+                            _id: item._id,
+                            type: item.type,
+                            max_items: item.max_items
+                        });
+                    }
+                });
+
+                if (scope.widget) {
+                    workspaces.getActive()
+                    .then(function(workspace) {
+                        var widgets =  angular.copy(workspace.widgets);
+                        _.each(widgets, function(widget) {
+                            if (scope.widget._id === widget._id && scope.widget.multiple_id === widget.multiple_id) {
+                                widget.configuration = {};
+                                widget.configuration.groups = groups;
+                                if (scope.widget.configuration.label) {
+                                    widget.configuration.label = scope.widget.configuration.label;
+                                }
+                            }
+                        });
+                        workspaces.save(workspace, {'widgets': widgets})
+                        .then(function() {
+                            scope.showGlobalSavedSearches = false;
+                            scope.onclose();
+                        });
+                    });
+                } else if (scope.settings && scope.settings.desk) {
+                    desks.save(scope.deskLookup[scope.settings.desk._id], {monitoring_settings: groups})
+                    .then(function() {
+                        WizardHandler.wizard('aggregatesettings').finish();
+                    });
+                } else {
+                    workspaces.getActiveId()
+                    .then(function(activeWorkspace) {
+                        if (activeWorkspace.type === 'workspace') {
+                            preferencesService.get(PREFERENCES_KEY)
+                            .then(function(preferences) {
+                                var updates = {};
+                                if (preferences) {
+                                    updates[PREFERENCES_KEY] = preferences;
+                                }
+                                updates[PREFERENCES_KEY][activeWorkspace.id] = {groups: groups};
+                                preferencesService.update(updates, PREFERENCES_KEY)
+                                .then(function() {
+                                    WizardHandler.wizard('aggregatesettings').finish();
+                                });
+                            });
+                        } else if (activeWorkspace.type === 'desk') {
+                            desks.save(scope.deskLookup[activeWorkspace.id], {monitoring_settings: groups})
+                            .then(function() {
+                                WizardHandler.wizard('aggregatesettings').finish();
+                            });
+                        }
+                    });
+                    scope.showGlobalSavedSearches = false;
+                    scope.onclose();
+                }
+            };
+        }
+    };
+}
