+++ scripts/superdesk-monitoring/directives/DeskNotifications.js
@@ -1,0 +1,84 @@
+/**
+ * Displays the notifications of the desk of a given stage
+ */
+DeskNotifications.$inject = ['desks', 'deskNotifications', 'authoringWorkspace', '$timeout'];
+export function DeskNotifications(desks, deskNotifications, authoringWorkspace, $timeout) {
+    return {
+        scope: {stage: '=stage'},
+        templateUrl: 'scripts/superdesk-monitoring/views/desk-notifications.html',
+        link: function(scope) {
+
+            function init() {
+                // Update the figures if there's a desk mention message
+                init_notifications();
+                if (scope.default_incoming) {
+                    scope.$on('desk:mention', function() {$timeout(reload, 5000, true);});
+                }
+            }
+
+            function init_notifications () {
+                scope.desk = desks.stageLookup[scope.stage].desk;
+                scope.notifications = deskNotifications.getNotifications(scope.desk);
+                scope.default_incoming = desks.stageLookup[scope.stage].default_incoming;
+                scope.notificationCount = deskNotifications.getUnreadCount(scope.desk) || 0;
+                scope.deskLookup = desks.deskLookup;
+                scope.stageLookup = desks.stageLookup;
+            }
+
+            function reload() {
+                deskNotifications.reload();
+                init_notifications();
+            }
+
+            /**
+             * Opens the story in the notification
+             * and updates the notification as read
+             *
+             * @param {object} notification The notification to be checked
+             */
+            scope.open = function(notification) {
+                authoringWorkspace.view(notification.item);
+            };
+
+            /**
+             * Updates the notification as read
+             *
+             * @param {object} notification The notification to be checked
+             */
+            scope.acknowledge = function(notification) {
+                deskNotifications.markAsRead(notification, scope.desk);
+                $timeout(reload, 5000);
+            };
+
+            function getRecipient(notification) {
+                return _.find(notification.recipients, {'desk_id': scope.desk});
+            }
+
+            /**
+             * Checks if the given notification is read
+             *
+             * @param {object} notification The notification to be checked
+             * @return {boolean} True if the notification is read by any user
+             */
+            scope.isRead = function(notification) {
+                var recipient = getRecipient(notification);
+                return recipient && recipient.read;
+            };
+
+            /**
+             * Returns the name of the user who read the notification
+             *
+             * @param {object} notification The notification to be checked
+             * @return {string} Display name of the user
+             */
+            scope.readBy = function(notification) {
+                var recipient = getRecipient(notification);
+                if (recipient && recipient.read) {
+                    return desks.userLookup[recipient.user_id].display_name;
+                }
+            };
+
+            init();
+        }
+    };
+}
