+++ scripts/superdesk-monitoring/directives/MonitoringGroupHeader.js
@@ -1,0 +1,5 @@
+export function MonitoringGroupHeader() {
+    return {
+        templateUrl: 'scripts/superdesk-monitoring/views/monitoring-group-header.html'
+    };
+}
