+++ scripts/superdesk-monitoring/directives/index.js
@@ -1,0 +1,8 @@
+export { MonitoringView } from './MonitoringView';
+export { MonitoringGroup } from './MonitoringGroup';
+export { MonitoringGroupHeader } from './MonitoringGroupHeader';
+export { DeskNotifications } from './DeskNotifications';
+export { ItemActionsMenu } from './ItemActionsMenu';
+
+export { AggregateSettings } from './AggregateSettings';
+export { SortGroups } from './SortGroups';
