+++ scripts/superdesk-monitoring/filters/index.js
@@ -1,0 +1,15 @@
+export function SplitFilter() {
+    return function(input) {
+        input = input || '';
+        var out = '';
+        for (var i = 0; i < input.length; i++) {
+            if (input.charAt(i) === input.charAt(i).toUpperCase()) {
+                out = out + ' ';
+            }
+
+            out = out + input.charAt(i);
+        }
+
+        return out;
+    };
+}
