+++ scripts/superdesk-monitoring/services/index.js
@@ -1,0 +1,1 @@
+export { CardsService } from './CardsService';
