+++ scripts/superdesk-monitoring/tests/monitoring.spec.js
